+++ JSM.Insight/Public/Get-InsightStatusTypes.ps1
@@ -2,6 +2,7 @@
 function Get-InsightStatusTypes {
     [CmdletBinding()]
     param (
+        [string]$objectSchemaId,
         [String]$Version = "1",
         [string]$InsightCreds = $InsightCreds,
         [string]$InsightWorkspaceID = $InsightWorkspaceID
@@ -13,20 +14,15 @@ function Get-InsightStatusTypes {
     }
     
     process {
-        $Request = [System.UriBuilder]"https://api.atlassian.com/jsm/insight/workspace/$InsightWorkspaceID/v$Version/config/statustype/"
+        if ($objectSchemaId) {
+            $Request = [System.UriBuilder]"https://api.atlassian.com/jsm/insight/workspace/$InsightWorkspaceID/v$Version/config/statustype/?objectSchemaId=$objectSchemaId"
+        }
+        else {
+            $Request = [System.UriBuilder]"https://api.atlassian.com/jsm/insight/workspace/$InsightWorkspaceID/v$Version/config/statustype/"
+        }
     }
     
     end {
-        try {
-            $response = Invoke-RestMethod -Uri $Request.Uri -Headers $headers -Method GET
-        }
-        catch {
-            Write-Verbose "[$($MyInvocation.MyCommand.Name)] Failed"
-            Write-Error -Message "$($_.Exception.Message)" -ErrorId $_.Exception.Code -Category InvalidOperation
-        } 
-
-        $response
-
-        Write-Verbose "[$($MyInvocation.MyCommand.Name)] Complete"
+        Invoke-RestMethod -Uri $Request.Uri -Headers $headers -Method GET
     }
 }